#6 extracted translations
