+++ src/js/helpers/index.js
@@ -1,3 +1,3 @@
 export * from './history';
 export * from './urlHelper';
-export * from './tableHelpers';
+export * from './utils';
