--- src/js/helpers/tableHelpers.js
@@ -1,7 +1,0 @@
-export const tableHelpers = {
-    filterByName
-};
-
-function filterByName(searchText, data) {
-    return data.filter(item => item.name.toLowerCase().includes(searchText.toLowerCase()));
-}
