+++ src/js/translations/de.js
@@ -1,0 +1,34 @@
+export const translations = {
+    attachments: 'Anhänge',
+    category: 'Kategorie',
+    create_method: 'Methode erstellen',
+    explanation_placeholder: 'Hier könnte eine Erklärung stehen.',
+    level: 'Level',
+    method_level_select_placeholder: 'Methoden-Level auswählen...',
+    method_name: 'Methodenname',
+    method_type_select_placeholder: 'Methodentyp auswählen...',
+    name: 'Name',
+    next: 'Weiter',
+    no_method_levels_found: 'Es existieren keine Methoden-Level',
+    no_method_types_found: 'Es existieren keine Methodentypen',
+    no_seminar_goals_found: 'Es existieren keine Seminarziele für das ausgwählte Seminar',
+    no_seminars_found: 'Es existieren keine Seminare',
+    previous: 'Zurück',
+    required: 'Verpflichtend',
+    role_type: 'Rollen Typ',
+    save: 'Speichern',
+    search_prefix: 'Suche',
+    section: 'Sektion',
+    seminar: 'Seminar',
+    seminar_goals: 'Seminarziele',
+    seminar_goals_select_placeholder: 'Seminarziele auswählen...',
+    seminar_roles: 'Seminarrollen',
+    seminar_select_placeholder: 'Seminar auswählen...',
+    seminar_types: 'Seminartypen',
+    step_attachments: 'Anänge',
+    step_content: 'Inhalt',
+    step_metadata: 'Metadaten',
+    step_summary: 'Zusammenfassung',
+    type: 'Typ',
+    unknown: 'Unbekannt',
+};
