+++ src/js/translations/index.js
@@ -1,0 +1,1 @@
+export * from './de';
